@@ -49,7 +49,7 @@
   },
   "gitHead": "2513190cc1de00d1049e35f9c942b779447530f9",
   "engines": {
-    "node": ">=18.0.0"
+    "node": ">=21.0.0"
   },
   "engineStrict": true,
   "dependencies": {
