@@ -1,8 +1,5 @@
 import { WikiHistoryPageScraper } from './scrapers/wiki-history-scraper.js';
-import { createRequire } from 'module';
-const require = createRequire(import.meta.url);
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const packageJson = require('../package.json');
+import packageJson from '../package.json' with { type: "json" };
 import { scrapeAndCollect } from './scrapers/collector.js';
 import { dateToFilename } from './utils/filesystem.js';
 import { Command } from 'commander';
